chore: separate out build tsconfig
