+++ packages/entity-codemod/tsconfig.build.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../tsconfig.build.json",
+  "exclude": ["**/__testfixtures__/**"]
+}
