+++ packages/entity-codemod/tsconfig.json
@@ -1,8 +1,5 @@
 {
   "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "build",
-  },
   "include": ["src"],
-  "exclude": ["**/__testfixtures__/**"]
+  "exclude": ["**/__testfixtures__/**"],
 }
