+++ packages/entity-database-adapter-knex/tsconfig.build.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../tsconfig.build.json",
+  "include": ["src"],
+}
