+++ packages/entity-example/tsconfig.json
@@ -1,7 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "noEmit": true,
-  },
   "include": ["src"]
 }
