+++ packages/entity-full-integration-tests/tsconfig.build.json
@@ -1,6 +1,6 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../../tsconfig.build.json",
   "compilerOptions": {
     "noEmit": true,
-  },
+  }
 }
