--- packages/entity-ip-address-field/README.md
@@ -1,3 +1,0 @@
-# @expo/entity-ip-address-field
-
-A package containing the IP address EntityField definitions for @expo/entity. Supports addresses with either IPv4 or IPv6 syntax.
