+++ packages/entity-secondary-cache-redis/tsconfig.json
@@ -1,7 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "build",
-  },
   "include": ["src"]
 }
