+++ packages/entity/package.json
@@ -10,6 +10,7 @@
   "types": "build/index.d.ts",
   "scripts": {
     "tsc": "tsc",
+    "build": "tsc -b tsconfig.build.json",
     "clean": "rm -rf build coverage coverage-integration",
     "lint": "eslint src",
     "lint-fix": "eslint src --fix",
