+++ packages/entity/src/internal/__tests__/ReadThroughEntityCache-test.ts
@@ -3,13 +3,16 @@ import { verify, mock, instance, when, anything } from 'ts-mockito';
 import EntityConfiguration from '../../EntityConfiguration';
 import { UUIDField } from '../../EntityFields';
 import IEntityCacheAdapter from '../../IEntityCacheAdapter';
+import {
+  deepEqualEntityAware,
+  isEqualWithEntityAware,
+} from '../../utils/__testfixtures__/TSMockitoExtensions';
 import ReadThroughEntityCache, { CacheStatus } from '../ReadThroughEntityCache';
 import {
   SingleFieldHolder,
   SingleFieldValueHolder,
   SingleFieldValueHolderMap,
 } from '../SingleFieldHolder';
-import { deepEqualEntityAware, isEqualWithEntityAware } from './TSMockitoExtensions';
 
 type BlahFields = {
   id: string;
