+++ packages/entity/src/utils/__testfixtures__/TSMockitoExtensions.ts
@@ -1,9 +1,12 @@
 import isEqualWith from 'lodash/isEqualWith';
 import { Matcher } from 'ts-mockito/lib/matcher/type/Matcher';
 
-import { SerializableKeyMap } from '../../utils/collections/SerializableKeyMap';
-import { CompositeFieldHolder, CompositeFieldValueHolder } from '../CompositeFieldHolder';
-import { SingleFieldHolder, SingleFieldValueHolder } from '../SingleFieldHolder';
+import {
+  CompositeFieldHolder,
+  CompositeFieldValueHolder,
+} from '../../internal/CompositeFieldHolder';
+import { SingleFieldHolder, SingleFieldValueHolder } from '../../internal/SingleFieldHolder';
+import { SerializableKeyMap } from '../collections/SerializableKeyMap';
 
 export function isEqualWithEntityAware(expected: any, actual: any): boolean {
   return isEqualWith(expected, actual, (expected: any, actual: any): boolean | undefined => {
