+++ tsconfig.build.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": [
+    "**/__tests__/**",
+    "**/__integration-tests__/**",
+    "**/__mocks__/**",
+    "**/__testfixtures__/**"
+  ],
+}
